invoice

+++ src/main/java/com/app/service/InvoiceService.java
@@ -1,9 +1,8 @@
 package com.app.service;
 
 import com.app.dto.InvoiceDto;
-import com.app.entity.Invoice;
 
 public interface InvoiceService {
 
-    InvoiceDto findInvoiceById(Integer id);
+    InvoiceDto findInvoiceById(Long id);
 }
